+++ packages/learn-card-types/src/vc.ts
@@ -19,14 +19,14 @@ export const ImageValidator = z.string().or(
 export type Image = z.infer<typeof ImageValidator>;
 
 export const GeoCoordinatesValidator = z.object({
-    type: z.string().array().nonempty(),
+    type: z.string().min(1).or(z.string().array().nonempty()),
     latitude: z.number(),
     longitude: z.number(),
 });
 export type GeoCoordinates = z.infer<typeof GeoCoordinatesValidator>;
 
 export const AddressValidator = z.object({
-    type: z.string().array().nonempty(),
+    type: z.string().min(1).or(z.string().array().nonempty()),
     addressCountry: z.string().optional(),
     addressCountryCode: z.string().optional(),
     addressRegion: z.string().optional(),
@@ -63,7 +63,7 @@ export const IdentifierTypeValidator = z
 export type IdentifierType = z.infer<typeof IdentifierTypeValidator>;
 
 export const IdentifierEntryValidator = z.object({
-    type: z.string().array().nonempty(),
+    type: z.string().min(1).or(z.string().array().nonempty()),
     identifier: z.string(),
     identifierType: IdentifierTypeValidator,
 });