+++ packages/learn-card-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @learncard/cli
 
+## 1.4.3
+
+### Patch Changes
+
+-   Updated dependencies [[`c1befdc`](https://github.com/learningeconomy/LearnCard/commit/c1befdc8a30d3cc111d938c530493b1a5b87aa00), [`c1befdc`](https://github.com/learningeconomy/LearnCard/commit/c1befdc8a30d3cc111d938c530493b1a5b87aa00)]:
+    -   @learncard/core@6.1.0
+    -   @learncard/types@3.0.0
+
 ## 1.4.2
 
 ### Patch Changes